Extract facet navigation to hook
